// test-app/src/app/api/swapi/people.ts
export interface People {

  birth_year: string;
  eye_color: string;
  gender: string;
  hair_color: string;
  height: number;
  homeworld: string;
  mass: number;
  name: string;
  skin_color: string;
  url: string;

  films: string[];
  species: string[];
  starships: string[];
  vehicles: string[];

  created: Date;
  edited: Date;

}