// test-app/src/app/api/swapi/species.ts
export interface Species {

  average_height: number;
  average_lifespan: number;
  classification: string;
  designation: string;
  eye_colors: string;
  hair_colors: string;
  homeworld: string;
  language: string;
  name: string;
  skin_colors: string;
  url: string;

  people: string[];
  films: string[];

  created: Date;
  edited: Date;
}